minor update at schemas, service, routers

+++ service/access_service.py
@@ -56,21 +56,21 @@ class AccessesService:
         if access.status != AccessStatus.ACTIVE:
             raise HTTPException(status_code=400, detail=f"Статус доступа не активен, внести изменения невозможно")
 
+        if update_data.status is not None:
+            if update_data.status == AccessStatus.REVOKED:
+                access.expires_at = now
+            if access.status == AccessStatus.ACTIVE and update_data.status == AccessStatus.EXPIRED:
+                if update_data.expires_at > now:
+                    raise HTTPException(status_code=400, detail="Текущий статус нельзя перевести в истекший")
+            if access.status == AccessStatus.EXPIRED and update_data.status == AccessStatus.ACTIVE:
+                raise HTTPException(status_code=400, detail="Нельзя активировать истекший доступ")
+            if access.status == AccessStatus.REVOKED and update_data.status == AccessStatus.ACTIVE:
+                raise HTTPException(status_code=400, detail="Текущий статус нельзя перевести в активный")
+            if access.status == AccessStatus.EXPIRED and update_data.status == AccessStatus.REVOKED:
+                raise HTTPException(status_code=400, detail="Текущий статус нельзя перевести в отозван")
+            access.status = update_data.status
 
-        if update_data.status == AccessStatus.REVOKED:
-            access.expires_at = now
-        if access.status == AccessStatus.ACTIVE and update_data.status == AccessStatus.EXPIRED:
-            if update_data.expires_at > now:
-                raise HTTPException(status_code=400, detail="Текущий статус нельзя перевести в истекший")
-        if access.status == AccessStatus.EXPIRED and update_data.status == AccessStatus.ACTIVE:
-            raise HTTPException(status_code=400, detail="Нельзя активировать истекший доступ")
-        if access.status == AccessStatus.REVOKED and update_data.status == AccessStatus.ACTIVE:
-            raise HTTPException(status_code=400, detail="Текущий статус нельзя перевести в активный")
-        if access.status == AccessStatus.EXPIRED and update_data.status == AccessStatus.REVOKED:
-            raise HTTPException(status_code=400, detail="Текущий статус нельзя перевести в отозван")
-
-
-        if update_data.expires_at:
+        if update_data.expires_at is not None:
             if update_data.expires_at.tzinfo is None:
                 update_data.expires_at = update_data.expires_at.replace(tzinfo=timezone.utc)
             if update_data.expires_at < now and update_data.status == AccessStatus.ACTIVE:
@@ -80,11 +80,10 @@ class AccessesService:
                 )
             if update_data.expires_at.tzinfo is None:
                 update_data.expires_at = update_data.expires_at.replace(tzinfo=timezone.utc)
+            access.expires_at = update_data.expires_at
 
-
-        access.expires_at = update_data.expires_at
-        access.status = update_data.status
-        access.comment = update_data.comment
+        if update_data.comment is not None:
+            access.comment = update_data.comment
 
         self.db.commit()
         self.db.refresh(access)
@@ -95,6 +94,8 @@ class AccessesService:
 
         if access is None:
             raise HTTPException(status_code=404, detail="Доступ с указанным ID не найден")
+        if access.status == AccessStatus.ACTIVE:
+            raise HTTPException(status_code=409, detail='Нельзя удалить активный доступ')
 
         self.repo.delete(access)
         self.db.commit()
